Replace EmptyListCell with label
