+++ Cineaste/CoreDataHelper/FetchedResultsManager.swift
@@ -51,25 +51,13 @@ extension FetchedResultsManager: NSFetchedResultsControllerDelegate {
     func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
         switch type {
         case .insert:
-            guard let indexPath = newIndexPath else { return }
-            if controller.fetchedObjects?.count == 1 {
-                // First object inserted, "empty cell" is replaced by "object cell"
-                delegate?.updateRows(at: [indexPath])
-            } else {
-                delegate?.insertRows(at: [indexPath])
-            }
+        guard let indexPath = newIndexPath else { return }
+            delegate?.insertRows(at: [indexPath])
         case .delete:
-            guard let indexPath = indexPath,
-                let fetchedObjects = controller.fetchedObjects
-                else { return }
-            if fetchedObjects.isEmpty {
-                // Last object removed, "object cell" is replaced by "empty cell"
-                delegate?.updateRows(at: [indexPath])
-            } else {
-                delegate?.deleteRows(at: [indexPath])
-            }
-        case .update:
             guard let indexPath = indexPath else { return }
+            delegate?.deleteRows(at: [indexPath])
+        case .update:
+        guard let indexPath = newIndexPath else { return }
             delegate?.updateRows(at: [indexPath])
         case .move:
             guard let indexPath = indexPath, let newIndexPath = newIndexPath else { return }
